% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spain_regional_cases.R
\name{get_spain_regional_cases}
\alias{get_spain_regional_cases}
\title{Spain regional cases daily}
\usage{
get_spain_regional_cases()
}
\value{
A dataframe of daily Spanish regional case counts.
}
\description{
Extract regional spanish case counts.
\href{https://www.mscbs.gob.es/profesionales/saludPublica/ccayes/alertasActual/nCov-China/situacionActual.html}{Source}
}
\examples{

\dontrun{

# Regional map

library(sf)
library(rnaturalearth)
library(ggplot2)

data <- get_spain_regional_cases()

regions <- rnaturalearth::ne_states("Spain", returnclass = "sf")

regions_with_data <- regions \%>\%
  group_by(region) \%>\%
  summarise(subregions = n(), do_union = TRUE) \%>\%
  mutate(region = str_replace_all(region, "í", "i"),
         region = str_replace_all(region, "ó", "o"),
         region = str_replace_all(region, "ñ", "n"),
         region = str_replace_all(region, "Foral de Navarra", "Navarra"),
         region = str_replace_all(region, "Valenciana", "C. Valenciana"),
         region = str_replace_all(region, "Canary Is\\\\.", "Canarias"),
         region = str_replace_all(region, "Islas Baleares", "Baleares"),
         region = str_replace_all(region, "-", " ")
         ) \%>\%
  left_join(data, by = c("region" = "region"))

# Map: Spain mainland + Baleares
es_mainland <- ggplot2::ggplot(regions_with_data) +
  geom_sf(ggplot2::aes(fill = cases)) +
  coord_sf(crs = st_crs(4326), xlim = c(-11, 4), ylim = c(35, 44))+
  theme_bw()

# Map inset: Canarias
es_canarias <- dplyr::filter(regions_with_data, region == "Canarias")
es_canarias <- ggplot2::ggplot(es_canarias) +
  geom_sf(aes(fill = cases)) +  coord_sf(datum = NA) +
  xlab(es_canarias$region) +
  theme_bw() +  theme(legend.position = "none")

# Map: mainland with insets
es_mainland +
  annotation_custom(
    grob = ggplotGrob(es_canarias),  xmin = -11,  xmax = -7,  ymin = 33.5,  ymax = 37.5)


}
##Code
get_spain_regional_cases
}
